backend/requirements.txt
asgiref==3.11.1
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.14
gunicorn==25.1.0
inflection==0.5.1
joblib==1.5.3
numpy==1.26.2
packaging==26.0
Pillow==10.0.0
PyJWT==2.11.0
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.3
reportlab==4.0.7
scikit-learn==1.3.2
scipy==1.11.4
sqlparse==0.5.5
threadpoolctl==3.6.0
tzdata==2025.3
uritemplate==4.2.0
dj-database-url
whitenoise
psycopg2-binary
